+++ RandomSkunk.Results/Maybe{T}.cs
@@ -173,7 +173,9 @@ public partial struct Maybe<T> : IEquatable<Maybe<T>>
     /// Creates a <c>Fail</c> result.
     /// </summary>
     /// <param name="errorMessage">The error message.</param>
-    /// <param name="stackTrace">The optional stack trace.</param>
+    /// <param name="stackTrace">
+    /// The optional stack trace. If <see langword="null"/>, then a generated stack trace is used.
+    /// </param>
     /// <param name="errorCode">The optional error code.</param>
     /// <param name="errorIdentifier">The optional identifier of the error.</param>
     /// <param name="errorType">
@@ -194,7 +196,7 @@ public partial struct Maybe<T> : IEquatable<Maybe<T>>
         Error? innerError = null) =>
         Fail(new Error(errorMessage, errorType)
         {
-            StackTrace = stackTrace,
+            StackTrace = stackTrace ?? new StackTrace(1).ToString(),
             ErrorCode = errorCode,
             Identifier = errorIdentifier,
             InnerError = innerError,