+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- Add generated stack traces to additional fail factory methods.
+
 ## [1.0.0-alpha13] - 2022-06-03
 
 ### Added